+++ pkg/services/mssqlfg/database_pair_fe_primary_deprovision.go
@@ -16,12 +16,16 @@ func (d *databasePairFePrimaryManager) GetDeprovisioner(
 			d.deletePriARMDeployment,
 		),
 		service.NewDeprovisioningStep(
-			"deleteSecARMDeployment",
-			d.deleteSecARMDeployment,
-		),
-		service.NewDeprovisioningStep(
 			"deleteFgARMDeployment",
 			d.deleteFgARMDeployment,
+		),
+		service.NewDeprovisioningStep(
+			"deletePriARMDeployment",
+			d.deletePriARMDeployment,
+		),
+		service.NewDeprovisioningStep(
+			"deleteSecARMDeployment",
+			d.deleteSecARMDeployment,
 		),
 		service.NewDeprovisioningStep(
 			"deletePriDatabase",
@@ -43,11 +47,11 @@ func (d *databasePairFePrimaryManager) deletePriARMDeployment(
 	instance service.Instance,
 ) (service.InstanceDetails, error) {
 	dt := instance.Details.(*databasePairInstanceDetails)
-	err := d.armDeployer.Delete(
+	ppp := instance.Parent.ProvisioningParameters
+	if err := d.armDeployer.Delete(
 		dt.PriARMDeploymentName,
-		instance.Parent.ProvisioningParameters.GetString("primaryResourceGroup"),
-	)
-	if err != nil {
+		ppp.GetString("primaryResourceGroup"),
+	); err != nil {
 		return nil, fmt.Errorf("error deleting ARM deployment: %s", err)
 	}
 	return instance.Details, nil
@@ -58,11 +62,11 @@ func (d *databasePairFePrimaryManager) deleteSecARMDeployment(
 	instance service.Instance,
 ) (service.InstanceDetails, error) {
 	dt := instance.Details.(*databasePairInstanceDetails)
-	err := d.armDeployer.Delete(
+	ppp := instance.Parent.ProvisioningParameters
+	if err := d.armDeployer.Delete(
 		dt.SecARMDeploymentName,
-		instance.Parent.ProvisioningParameters.GetString("secResourceGroup"),
-	)
-	if err != nil {
+		ppp.GetString("secondaryResourceGroup"),
+	); err != nil {
 		return nil, fmt.Errorf("error deleting ARM deployment: %s", err)
 	}
 	return instance.Details, nil
@@ -73,11 +77,11 @@ func (d *databasePairFePrimaryManager) deleteFgARMDeployment(
 	instance service.Instance,
 ) (service.InstanceDetails, error) {
 	dt := instance.Details.(*databasePairInstanceDetails)
-	err := d.armDeployer.Delete(
+	ppp := instance.Parent.ProvisioningParameters
+	if err := d.armDeployer.Delete(
 		dt.FgARMDeploymentName,
-		instance.Parent.ProvisioningParameters.GetString("primaryResourceGroup"),
-	)
-	if err != nil {
+		ppp.GetString("primaryResourceGroup"),
+	); err != nil {
 		return nil, fmt.Errorf("error deleting ARM deployment: %s", err)
 	}
 	return instance.Details, nil
@@ -91,13 +95,16 @@ func (d *databasePairFePrimaryManager) deletePriDatabase(
 	defer cancel()
 	dt := instance.Details.(*databasePairInstanceDetails)
 	pdt := instance.Parent.Details.(*dbmsPairInstanceDetails)
-	if _, err := d.databasesClient.Delete(
-		ctx,
-		instance.Parent.ProvisioningParameters.GetString("primaryResourceGroup"),
-		pdt.PriServerName,
-		dt.DatabaseName,
-	); err != nil {
-		return nil, fmt.Errorf("error deleting sql database: %s", err)
+	ppp := instance.Parent.ProvisioningParameters
+	if dt.DatabaseName != "" {
+		if _, err := d.databasesClient.Delete(
+			ctx,
+			ppp.GetString("primaryResourceGroup"),
+			pdt.PriServerName,
+			dt.DatabaseName,
+		); err != nil {
+			return nil, fmt.Errorf("error deleting sql database: %s", err)
+		}
 	}
 	return instance.Details, nil
 }
@@ -110,13 +117,16 @@ func (d *databasePairFePrimaryManager) deleteSecDatabase(
 	defer cancel()
 	dt := instance.Details.(*databasePairInstanceDetails)
 	pdt := instance.Parent.Details.(*dbmsPairInstanceDetails)
-	if _, err := d.databasesClient.Delete(
-		ctx,
-		instance.Parent.ProvisioningParameters.GetString("secResourceGroup"),
-		pdt.SecServerName,
-		dt.DatabaseName,
-	); err != nil {
-		return nil, fmt.Errorf("error deleting sql database: %s", err)
+	ppp := instance.Parent.ProvisioningParameters
+	if dt.DatabaseName != "" {
+		if _, err := d.databasesClient.Delete(
+			ctx,
+			ppp.GetString("secondaryResourceGroup"),
+			pdt.SecServerName,
+			dt.DatabaseName,
+		); err != nil {
+			return nil, fmt.Errorf("error deleting sql database: %s", err)
+		}
 	}
 	return instance.Details, nil
 }
@@ -129,13 +139,16 @@ func (d *databasePairFePrimaryManager) deleteFailoverGroup(
 	defer cancel()
 	dt := instance.Details.(*databasePairInstanceDetails)
 	pdt := instance.Parent.Details.(*dbmsPairInstanceDetails)
-	if _, err := d.databasesClient.Delete(
-		ctx,
-		instance.Parent.ProvisioningParameters.GetString("primaryResourceGroup"),
-		pdt.PriServerName,
-		dt.FailoverGroupName,
-	); err != nil {
-		return nil, fmt.Errorf("error deleting failover group: %s", err)
+	ppp := instance.Parent.ProvisioningParameters
+	if dt.FailoverGroupName != "" {
+		if _, err := d.databasesClient.Delete(
+			ctx,
+			ppp.GetString("primaryResourceGroup"),
+			pdt.PriServerName,
+			dt.FailoverGroupName,
+		); err != nil {
+			return nil, fmt.Errorf("error deleting failover group: %s", err)
+		}
 	}
 	return instance.Details, nil
 }
